Maintain lazy import state when resolve fails
Make sure to leave both the module state and the sys.meta_path state as it was prior to calling resolve_lazy_imports for the case where the resolution fails, for example due to a network connectivity issue between grpc client and server.

While adding a test case for this, I noticed that the cleanup done by the lazy_cleaner fixture was not actually resetting state sufficiently to exercise subsequent test cases, so I fixed that, too. This should make all the lazy import test cases meaningful, whereas only the first one was before.

This should address issue #1.