Set timeouts on all reflection service queries
Timeout is set arbitrarily to 10 seconds, which should be sufficiently long to connote "not going to respond any time soon".

This exposed a deficiency in the grpc pytest plugin that doesn't support any of the optional args to grpc handler methods, which had to be hacked around in a pretty ugly way, but at least that enabled a hook for the tests to verify that all reflection queries have timeouts.